Year 2 Block D/pyproject.toml
[tool.poetry]
name         = "emotion_mvp"
version      = "0.1.0"
description  = "Multilingual → English emotion-classifier MVP"
authors      = [
  "Musaed Alfareh <225739@buas.nl>",
  "Michon Godijn  <232849@buas.nl>",
  "Edoardo Pierezza <231412@buas.nl>",
  "Rafał Stańczyszyn <234778@buas.nl>",
  "Christopher Schreuder <232202@buas.nl>"
]
license      = "MIT"
readme       = "README.md"
homepage     = "https://github.com/your-org/emotion_mvp"
repository   = "https://github.com/your-org/emotion_mvp"

[tool.poetry.dependencies]
python             = "^3.10"
numpy              = "<2"
transformers       = "^4.40"
torch              = "^2.2"
langdetect         = "^1.0"
click              = "^8.1"
fastapi            = "^0.111"
uvicorn            = "^0.29"
python-dotenv      = "^1.0"
requests           = "^2.32"
pandas             = "^2.2"
yt-dlp             = "^2025.4.30"
openai-whisper     = {git = "https://github.com/openai/whisper.git"}
tensorflow = { version="^2.12", markers="sys_platform != 'win32'" }

# Azure Blob Storage SDKs
azure-identity     = "^1.15.0"
azure-storage-blob = "^12.18.0"

[tool.poetry.group.dev.dependencies]
pytest        = "^8.2"
pytest-mock   = "^3.14"
ruff          = "^0.4"
sphinx        = "^7.3"
sphinx-click  = "^5.1"
pytest-cov = "^6.2.1"

[tool.poetry.scripts]
emotion-cli = "emotion_mvp.cli:main"

[build-system]
requires      = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = """
    --cov=emotion_mvp
    --cov-report=term-missing
    --cov-fail-under=85
"""
